Update to allow statsbeat to collect dynamic feature changes.
