+++ sdk/monitor/monitor-opentelemetry-exporter/src/export/statsbeat/longIntervalStatsbeatMetrics.ts
@@ -34,7 +34,6 @@ let instance: LongIntervalStatsbeatMetrics | null = null;
  * @internal
  */
 class LongIntervalStatsbeatMetrics extends StatsbeatMetrics {
-  private AZURE_MONITOR_STATSBEAT_FEATURES = process.env.AZURE_MONITOR_STATSBEAT_FEATURES;
   private statsCollectionLongInterval: number = 86400000; // 1 day
   // Custom dimensions
   private cikey: string;
@@ -70,16 +69,7 @@ class LongIntervalStatsbeatMetrics extends StatsbeatMetrics {
       connectionString: this.connectionString,
     };
 
-    if (this.AZURE_MONITOR_STATSBEAT_FEATURES) {
-      try {
-        this.feature = JSON.parse(this.AZURE_MONITOR_STATSBEAT_FEATURES).feature;
-        this.instrumentation = JSON.parse(this.AZURE_MONITOR_STATSBEAT_FEATURES).instrumentation;
-      } catch (error: any) {
-        diag.error(
-          `LongIntervalStatsbeat: Failed to parse features/instrumentations (error ${error})`,
-        );
-      }
-    }
+    this.setFeatures();
 
     this.longIntervalStatsbeatMeterProvider = new MeterProvider();
     this.longIntervalAzureExporter = new AzureMonitorStatsbeatExporter(exporterConfig);
@@ -155,6 +145,7 @@ class LongIntervalStatsbeatMetrics extends StatsbeatMetrics {
   }
 
   private featureCallback(observableResult: BatchObservableResult) {
+    this.setFeatures();
     let attributes;
     if (this.instrumentation) {
       attributes = {
@@ -172,6 +163,20 @@ class LongIntervalStatsbeatMetrics extends StatsbeatMetrics {
         type: StatsbeatFeatureType.FEATURE,
       };
       observableResult.observe(this.featureStatsbeatGauge, 1, { ...attributes });
+    }
+  }
+
+  private setFeatures() {
+    let statsbeatFeatures = process.env.AZURE_MONITOR_STATSBEAT_FEATURES;
+    if (statsbeatFeatures) {
+      try {
+        this.feature = JSON.parse(statsbeatFeatures).feature;
+        this.instrumentation = JSON.parse(statsbeatFeatures).instrumentation;
+      } catch (error: any) {
+        diag.error(
+          `LongIntervalStatsbeat: Failed to parse features/instrumentations (error ${error})`,
+        );
+      }
     }
   }
